% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{calc_R2}
\alias{calc_R2}
\title{Function for calculating coefficient of determination#' Read plink file through R}
\usage{
calc_R2(y, y_pred)
}
\arguments{
\item{y}{observation vector}

\item{y_pred}{prediction vector}
}
\description{
Function for calculating coefficient of determination#' Read plink file through R
}
\examples{
calc_R2()
}
\keyword{R^2}
\keyword{calcaulate}
