% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{constant_process}
\alias{constant_process}
\title{Function for mapping snps to genes in 3D method (e.g. Domain, Loop, TAD)}
\usage{
constant_process(geno, expression, chr, window)
}
\arguments{
\item{geno}{genotype dataframe}

\item{expression}{expression dataframe}

\item{chr}{chromosome number}

\item{window}{+/- (window) kb to consider around each gene}
}
\description{
Function for mapping snps to genes in 3D method (e.g. Domain, Loop, TAD)
}
\examples{
constant_process()
}
\keyword{constant}
\keyword{method}
