% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{elnet}
\alias{elnet}
\title{Cross-validation}
\usage{
elnet(
  geno_tuning,
  expression_tuning,
  geno_testing,
  expression_testing,
  penalty_k,
  lambda_list
)
}
\arguments{
\item{geno_tuning}{tuning genotype set}

\item{expression_tuning}{tuning expression set}

\item{geno_testing}{testing genotype set}

\item{expression_testing}{testing expression set}

\item{penalty_k}{penalty factor}

\item{lambda_list}{list of l1 penalty value}
}
\description{
Cross-validation
}
\examples{
elnet()
}
\keyword{cross-validation}
