% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{expression_scale}
\alias{expression_scale}
\title{Function for scaling expression dataframe to mean of zero}
\usage{
expression_scale(expression_subset, sample_tuning, sample_testing)
}
\arguments{
\item{expression_subset}{input expression}

\item{sample_tuning}{index of tunning sample}

\item{sample_testing}{index of testing sample}
}
\description{
Function for scaling expression dataframe to mean of zero
}
\examples{
expression_scale()
}
\keyword{Scaling}
\keyword{expression}
