% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{file_helper}
\alias{file_helper}
\title{Function for splitting jobs}
\usage{
file_helper(x, n)
}
\arguments{
\item{x}{joblist}

\item{n}{number of total file}
}
\description{
Function for splitting jobs
}
\examples{
file_helper()
}
\keyword{job}
\keyword{split}
