% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{geno_coord}
\alias{geno_coord}
\title{Function for filling in snp coordinates}
\usage{
geno_coord(geno)
}
\arguments{
\item{geno}{genotype dataframe}
}
\description{
Function for filling in snp coordinates
}
\examples{
geno_coord()
}
\keyword{constant}
\keyword{method}
