% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{geno_fill}
\alias{geno_fill}
\title{Function for filling in missing genotype}
\usage{
geno_fill(geno, type)
}
\arguments{
\item{geno}{genotype dataframe}

\item{type}{filling missing value}
}
\description{
Function for filling in missing genotype
}
\examples{
geno_fill()
}
\keyword{constant}
\keyword{method}
