% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{geno_scale}
\alias{geno_scale}
\title{Function for scaling genotype dataframe to mean of zero and sd of one}
\usage{
geno_scale(geno_subset, sample_tuning, sample_testing)
}
\arguments{
\item{geno_subset}{input genotype matrix}

\item{sample_tuning}{index of tunning sample}

\item{sample_testing}{index of testing sample}
}
\description{
Function for scaling genotype dataframe to mean of zero and sd of one
}
\examples{
geno_scale()
}
\keyword{Scaling}
\keyword{genotype}
