% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{pchic_process}
\alias{pchic_process}
\title{Function for mapping snps to genes in 3D method pchic}
\usage{
pchic_process(geno, expression, func)
}
\arguments{
\item{geno}{genotype dataframe}

\item{expression}{expression dataframe}

\item{func}{3D map dataframe}
}
\description{
Function for mapping snps to genes in 3D method pchic
}
\examples{
pchic_process()
}
\keyword{3D}
\keyword{method}
\keyword{pchic}
