% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{PUMICE.compute_weight}
\alias{PUMICE.compute_weight}
\title{Create gene expression prediction model using window type and penalty factor derived from the CV step}
\usage{
PUMICE.compute_weight(
  geno_path,
  exp_path,
  encode_path,
  output_path,
  window_path_list
)
}
\arguments{
\item{geno_path}{genotype path, plink format}

\item{exp_path}{expression path, txt format}

\item{encode_path}{epigenomic annotation path, txt format}

\item{output_path}{dir of your output file}

\item{window_path_list}{A list of 3d genomic winodows you need}
}
\description{
Create gene expression prediction model using window type and penalty factor derived from the CV step
}
\examples{
PUMICE.compute_weight()
}
\keyword{weights}
