% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.nested_cv.R
\name{PUMICE.nested_cv}
\alias{PUMICE.nested_cv}
\title{Run nested cross-validation to determine which window type and penalty factor are optimal}
\usage{
PUMICE.nested_cv(
  geno_path,
  exp_path,
  encode_path,
  output_path,
  window,
  window_path_list = NULL
)
}
\arguments{
\item{geno_path}{genotype path, plink format}

\item{exp_path}{expression path, txt format}

\item{encode_path}{epigenomic annotation path, txt format}

\item{output_path}{dir of your output file}

\item{window}{specify the constant or 3D window you need}

\item{window_path_list}{A list of 3d genomic winodows you need}
}
\description{
Run nested cross-validation to determine which window type and penalty factor are optimal
}
\examples{
PUMICE.nested_cv()
}
\keyword{nested.cv}
