% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PUMICE.compute_weights.R
\name{read_plink_custom}
\alias{read_plink_custom}
\title{Read plink file through R}
\usage{
read_plink_custom(root, impute = c("none", "avg", "random"))
}
\arguments{
\item{root}{plink file}

\item{impute}{How to impute missing value}
}
\description{
Read plink file through R
}
\examples{
read_plink_custom()
}
\keyword{plink}
\keyword{read}
